\name{ple2000cs}
\alias{ple2000cs}
\alias{ple2000cl}
\docType{data}
\title{IMARES plaice data}
\description{
To be added
}
\usage{
data(ple2000cs)
data(ple2000cl)

}
\format{
Formal class 'ceData' [package "COSTcore"] objects with 2 slots:
\cr@desc: description
\cr@ce: data frame of 18 variables  
\cr see ceData for details of the variables  
}
\details{

To be added 

}
\section{Warning }{These are test data sets only and should not to be used or cited without prior permission.}

\source{
  IMARES 
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(ple2000cs)
}
\keyword{datasets}
