\name{SIBM2006cs}
\alias{SIBM2006cs}
\alias{SIBM2006ce}
\alias{SIBM2006cl}
\docType{data}
\title{SIBM mixed species data}
\description{
To be added
}
\usage{
data(SIBM2006cs)
data(SIBM2006cl)
data(SIBM2006ce)
}
\format{
SIMB2006cs: Formal class 'csData' [package "COSTcore"] objects with 6 slots:
\cr@desc: description
\cr@tr: data.frame of 16 variables 
\cr@hh: data.frame of 29 variables 
\cr@sl: data.frame of 17 variables
\cr@hl: data.frame of 16 variables
\cr@ca: data.frame of 31 variables
\cr see csData for details of the variables 

SIMB2006cl: Formal class 'clData' [package "COSTcore"] objects with 2 slots:
\cr@desc: description
\cr@cl: data frame of 22 variables  
\cr see clData for details of the variables  

SIMB2006ce: Formal class 'ceData' [package "COSTcore"] objects with 2 slots:
\cr@desc: description
\cr@ce: data frame of 18 variables  
\cr see ceData for details of the variables  
}
\details{

To be added 

}
\section{Warning }{These are test data sets only and should not to be used or cited without prior permission.}

\source{
  SIBM 
}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(SIBM2006cs)
}
\keyword{datasets}
